refactor: domain 아키텍쳐에서 layerd 아키텍쳐로 변경
- domain 디렉토리에 있는 걸 entity, repository 로 이동
