+++ src/main/java/our/yurivongella/instagramclone/controller/dto/comment/CommentResponseDto.java
@@ -6,8 +6,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import our.yurivongella.instagramclone.controller.dto.member.MemberDto;
-import our.yurivongella.instagramclone.domain.comment.Comment;
-import our.yurivongella.instagramclone.domain.member.Member;
+import our.yurivongella.instagramclone.entity.Comment;
+import our.yurivongella.instagramclone.entity.Member;
 
 import java.time.LocalDateTime;
