+++ src/main/java/our/yurivongella/instagramclone/controller/dto/member/MemberDto.java
@@ -3,7 +3,7 @@ package our.yurivongella.instagramclone.controller.dto.member;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import our.yurivongella.instagramclone.domain.member.Member;
+import our.yurivongella.instagramclone.entity.Member;
 
 @Getter
 @NoArgsConstructor
