+++ src/main/java/our/yurivongella/instagramclone/entity/Follow.java
@@ -1,15 +1,10 @@
-package our.yurivongella.instagramclone.domain.follow;
-
-import java.util.List;
+package our.yurivongella.instagramclone.entity;
 
 import javax.persistence.*;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.ToString;
-import our.yurivongella.instagramclone.domain.BaseEntity;
-import our.yurivongella.instagramclone.domain.member.Member;
 
 @Getter
 @Entity
