+++ src/main/java/our/yurivongella/instagramclone/entity/Post.java
@@ -1,4 +1,4 @@
-package our.yurivongella.instagramclone.domain.post;
+package our.yurivongella.instagramclone.entity;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -19,9 +19,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import our.yurivongella.instagramclone.domain.BaseEntity;
-import our.yurivongella.instagramclone.domain.comment.Comment;
-import our.yurivongella.instagramclone.domain.member.Member;
 import our.yurivongella.instagramclone.exception.CustomException;
 import our.yurivongella.instagramclone.exception.ErrorCode;
