+++ src/main/java/our/yurivongella/instagramclone/entity/RefreshToken.java
@@ -1,4 +1,4 @@
-package our.yurivongella.instagramclone.domain.refeshtoken;
+package our.yurivongella.instagramclone.entity;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,7 +6,6 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.Table;
