+++ src/main/java/our/yurivongella/instagramclone/repository/CommentRepository.java
@@ -1,12 +1,11 @@
-package our.yurivongella.instagramclone.domain.comment;
+package our.yurivongella.instagramclone.repository;
 
 import java.util.List;
 
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import our.yurivongella.instagramclone.domain.member.Member;
-import our.yurivongella.instagramclone.domain.post.Post;
+import our.yurivongella.instagramclone.entity.Post;
+import our.yurivongella.instagramclone.entity.Comment;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
     List<Comment> findTop3ByPostOrderByIdDesc(Post post);
