+++ src/main/java/our/yurivongella/instagramclone/repository/FollowRepository.java
@@ -1,10 +1,11 @@
-package our.yurivongella.instagramclone.domain.follow;
+package our.yurivongella.instagramclone.repository;
 
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import our.yurivongella.instagramclone.domain.member.Member;
+import our.yurivongella.instagramclone.entity.Member;
+import our.yurivongella.instagramclone.entity.Follow;
 
 public interface FollowRepository extends JpaRepository<Follow, Long> {
     List<Follow> findByToMember(Member member);
