+++ src/main/java/our/yurivongella/instagramclone/repository/MediaUrlRepository.java
@@ -1,6 +1,7 @@
-package our.yurivongella.instagramclone.domain.post;
+package our.yurivongella.instagramclone.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import our.yurivongella.instagramclone.entity.MediaUrl;
 
 public interface MediaUrlRepository extends JpaRepository<MediaUrl, Long> {
 }
