+++ src/main/java/our/yurivongella/instagramclone/repository/post/CustomPostRepository.java
@@ -1,7 +1,8 @@
-package our.yurivongella.instagramclone.domain.post;
+package our.yurivongella.instagramclone.repository.post;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
+import our.yurivongella.instagramclone.entity.Post;
 
 public interface CustomPostRepository {
     Slice<Post> findAllByJoinFollow(Long memberId, Pageable pageable);
