+++ src/main/java/our/yurivongella/instagramclone/service/AuthService.java
@@ -14,10 +14,10 @@ import our.yurivongella.instagramclone.controller.dto.auth.SigninReqDto;
 import our.yurivongella.instagramclone.controller.dto.auth.SignupReqDto;
 import our.yurivongella.instagramclone.controller.dto.auth.TokenResDto;
 import our.yurivongella.instagramclone.controller.dto.auth.TokenReqDto;
-import our.yurivongella.instagramclone.domain.member.Member;
-import our.yurivongella.instagramclone.domain.member.MemberRepository;
-import our.yurivongella.instagramclone.domain.refeshtoken.RefreshToken;
-import our.yurivongella.instagramclone.domain.refeshtoken.RefreshTokenRepository;
+import our.yurivongella.instagramclone.entity.Member;
+import our.yurivongella.instagramclone.repository.MemberRepository;
+import our.yurivongella.instagramclone.entity.RefreshToken;
+import our.yurivongella.instagramclone.repository.RefreshTokenRepository;
 import our.yurivongella.instagramclone.exception.CustomException;
 import our.yurivongella.instagramclone.exception.ErrorCode;
 import our.yurivongella.instagramclone.jwt.TokenProvider;
