+++ src/main/java/our/yurivongella/instagramclone/service/CommentService.java
@@ -14,14 +14,14 @@ import lombok.extern.slf4j.Slf4j;
 import our.yurivongella.instagramclone.controller.dto.comment.CommentCreateDto;
 import our.yurivongella.instagramclone.controller.dto.comment.ProcessStatus;
 import our.yurivongella.instagramclone.controller.dto.comment.CommentResponseDto;
-import our.yurivongella.instagramclone.domain.comment.Comment;
-import our.yurivongella.instagramclone.domain.comment.CommentLike;
-import our.yurivongella.instagramclone.domain.comment.CommentLikeRepository;
-import our.yurivongella.instagramclone.domain.comment.CommentRepository;
-import our.yurivongella.instagramclone.domain.member.Member;
-import our.yurivongella.instagramclone.domain.member.MemberRepository;
-import our.yurivongella.instagramclone.domain.post.Post;
-import our.yurivongella.instagramclone.domain.post.PostRepository;
+import our.yurivongella.instagramclone.entity.Comment;
+import our.yurivongella.instagramclone.entity.CommentLike;
+import our.yurivongella.instagramclone.repository.CommentLikeRepository;
+import our.yurivongella.instagramclone.repository.CommentRepository;
+import our.yurivongella.instagramclone.entity.Member;
+import our.yurivongella.instagramclone.repository.MemberRepository;
+import our.yurivongella.instagramclone.entity.Post;
+import our.yurivongella.instagramclone.repository.post.PostRepository;
 import our.yurivongella.instagramclone.exception.CustomException;
 import our.yurivongella.instagramclone.exception.ErrorCode;
 import our.yurivongella.instagramclone.util.SecurityUtil;
