+++ src/test/java/our/yurivongella/instagramclone/controller/dto/MemberResponseDtoTest.java
@@ -3,7 +3,7 @@ package our.yurivongella.instagramclone.controller.dto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import our.yurivongella.instagramclone.controller.dto.member.MemberResponseDto;
-import our.yurivongella.instagramclone.domain.member.Member;
+import our.yurivongella.instagramclone.entity.Member;
 
 import static org.assertj.core.api.Assertions.*;
